+++ test/Test/Meta/RandomEffects.hs
@@ -1,4 +1,4 @@
-module Test.Meta.CommonEffect where
+module Test.Meta.RandomEffects where
 
 import           Data.Maybe
 
@@ -13,7 +13,7 @@ import           TestHS
 
 import Data.Numerics
 import           Data.Meta.Effects
-import           Data.Meta.CommonEffect
+import           Data.Meta.RandomEffects
 
 {-fastTests :: [Test]-}
 {-fastTests = [ -}
@@ -23,12 +23,14 @@ import           Data.Meta.CommonEffect
 ioTests :: [IO Test]
 ioTests = [ smd
           , rr
+          , testor
           , testrd
+          , testtau
           ]
 
 smd :: IO Test
 smd = do
-  let name = "Common Effect meta-analysis on continuous SMD"
+  let name = "Random Effects meta-analysis on continuous SMD"
   let studiesFile = "test/continuous.csv"
   csvData <- B.readFile studiesFile
   let estudies = C.decodeByName csvData
@@ -37,10 +39,10 @@ smd = do
     Left err -> return $ testFailed name $ ("error parsing csv",err)
     Right (_, studies) -> do
       let emds = rights $ fmap standardizedMeanDifference $ V.toList studies 
-      let ce = commonEffect emds
+      let ce = summary $ randomEffects emds
       let (SMD e v) = ce
       let foundce = (mapEstimate (\c -> roundDouble c 4) (SMD e v))
-      let expected = SMD 0.4143 0.0041
+      let expected = SMD 0.3582 0.0111
       if  foundce == expected
         then
           return $ testPassed name $ "nice!!"
@@ -49,7 +51,7 @@ smd = do
 
 rr :: IO Test
 rr = do
-  let name = "Common Effect meta-analysis on binary RR"
+  let name = "Random Effects meta-analysis on binary RR"
   let studiesFile = "test/binary.csv"
   csvData <- B.readFile studiesFile
   let estudies = C.decodeByName csvData
@@ -58,9 +60,29 @@ rr = do
     Left err -> return $ testFailed name $ ("error parsing csv",err)
     Right (_, studies) -> do
       let emds = rights $ fmap riskRatio $ V.toList studies 
-      let ce = commonEffect emds
+      let ce = randomEffectsRR emds
       let foundce = (mapEstimate (\c -> roundDouble c 4) ce)
-      let expected = RR 0.5775 (CI 0.4511 0.7392)
+      let expected = RR 0.6395 (CI 0.4283 0.9548)
+      if  foundce == expected
+        then
+          return $ testPassed name $ "nice!!"
+        else
+          return $ testFailed name $ (show expected, show foundce)
+
+testor :: IO Test
+testor = do
+  let name = "Random Effects meta-analysis on binary OR"
+  let studiesFile = "test/binary.csv"
+  csvData <- B.readFile studiesFile
+  let estudies = C.decodeByName csvData
+               :: Either String (C.Header, V.Vector Study)
+  case estudies of
+    Left err -> return $ testFailed name $ ("error parsing csv",err)
+    Right (_, studies) -> do
+      let emds = rights $ fmap oddsRatio $ V.toList studies 
+      let ce = randomEffectsOR emds
+      let foundce = (mapEstimate (\c -> roundDouble c 4) ce)
+      let expected = OR 0.5676 (CI 0.3554 0.9065)
       if  foundce == expected
         then
           return $ testPassed name $ "nice!!"
@@ -69,7 +91,7 @@ rr = do
 
 testrd :: IO Test
 testrd = do
-  let name = "Common Effect meta-analysis on binary RD"
+  let name = "Random Effects meta-analysis on binary RD"
   let studiesFile = "test/binary.csv"
   csvData <- B.readFile studiesFile
   let estudies = C.decodeByName csvData
@@ -78,7 +100,7 @@ testrd = do
     Left err -> return $ testFailed name $ ("error parsing csv",err)
     Right (_, studies) -> do
       let emds = rights $ fmap riskDifference $ V.toList studies 
-      let ce = commonEffect emds
+      let ce = summary $ randomEffects emds
       let foundce = (mapEstimate (\c -> roundDouble c 4) ce)
       let expected = RD (-0.1119) $ roundDouble (ciToVariance (CI (-0.1499) (-0.0739))) 4
       if  foundce == expected
@@ -86,3 +108,24 @@ testrd = do
           return $ testPassed name $ "nice!!"
         else
           return $ testFailed name $ (show expected, show foundce)
+
+testtau :: IO Test
+testtau = do
+  let name = "Random Effects meta-analysis tau test"
+  let studiesFile = "test/continuous.csv"
+  csvData <- B.readFile studiesFile
+  let estudies = C.decodeByName csvData
+               :: Either String (C.Header, V.Vector Study)
+  case estudies of
+    Left err -> return $ testFailed name $ ("error parsing csv",err)
+    Right (_, studies) -> do
+      let emds = rights $ fmap standardizedMeanDifference $ V.toList studies 
+      let ce = randomEffects emds
+      let foundce = (mapEstimate (\c -> roundDouble c 6) (τsq ce))
+      let expected = TauSquare 0.037311 0.001761 (CI 0 0.131214360151206)
+      if  foundce == expected
+        then
+          return $ testPassed name $ "nice!!"
+        else
+          return $ testFailed name $ (show expected, show foundce)
+