+++ ChangeLog.md
@@ -1,3 +1,9 @@
 # Changelog for meta-analysis
 
+- 0.1.1.0
+  Random Effects
+
+- 0.1.0.0
+  Common Effect
+
 ## Unreleased changes